# Conflicts:
#	ScientificReport.BLL/Services/MembershipService.cs
#	ScientificReport/Migrations/20190529212024_Initial.Designer.cs
#	ScientificReport/Migrations/20190529212024_Initial.cs
#	ScientificReport/Migrations/ScientificReportDbContextModelSnapshot.cs
#	ScientificReport/Models/SeedData.cs
#	ScientificReport/Views/Membership/Index.cshtml

+++ ScientificReport/Controllers/ScientificInternshipController.cs
@@ -1,8 +1,12 @@
 using System;
+using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using ScientificReport.BLL.Interfaces;
+using ScientificReport.Controllers.Utils;
+using ScientificReport.DAL.Entities;
 using ScientificReport.DAL.Roles;
+using ScientificReport.DTO.Models;
 using ScientificReport.DTO.Models.ScientificInternship;
 
 namespace ScientificReport.Controllers
@@ -11,17 +15,25 @@ namespace ScientificReport.Controllers
 	public class ScientificInternshipController : Controller
 	{
 		private readonly IScientificInternshipService _scientificInternshipService;
+		private readonly IUserProfileService _userProfileService;
+		private readonly IDepartmentService _departmentService;
 
-		public ScientificInternshipController(IScientificInternshipService scientificInternshipService)
+		public ScientificInternshipController(
+			IScientificInternshipService scientificInternshipService,
+			IUserProfileService userProfileService,
+			IDepartmentService departmentService
+		)
 		{
 			_scientificInternshipService = scientificInternshipService;
+			_userProfileService = userProfileService;
+			_departmentService = departmentService;
 		}
 
 		// GET: ScientificInternship
 		public IActionResult Index(ScientificInternshipIndexModel model)
 		{
-			model.ScientificInternships = _scientificInternshipService.GetPage(model.CurrentPage, model.PageSize);
-			model.Count = _scientificInternshipService.GetCount();
+			model.ScientificInternships = _scientificInternshipService.GetPageByRole(model.CurrentPage, model.PageSize, User);
+			model.Count = _scientificInternshipService.GetCountByRole(User);
 			return View(model);
 		}
 
@@ -39,6 +51,11 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
+			if (!UserHasPermission(scientificInternship))
+			{
+				return Forbid();
+			}
+
 			return View(scientificInternship);
 		}
 
@@ -54,7 +71,13 @@ namespace ScientificReport.Controllers
 			{
 				return View(model);
 			}
+			
 			_scientificInternshipService.CreateItem(model);
+			_scientificInternshipService.AddUser(
+				_scientificInternshipService.Get(si =>
+					si.Contents == model.Contents && si.PlaceOfInternship == model.PlaceOfInternship),
+				_userProfileService.Get(User));
+			
 			return RedirectToAction(nameof(Index));
 		}
 
@@ -71,8 +94,17 @@ namespace ScientificReport.Controllers
 			{
 				return NotFound();
 			}
+			
+			if (!UserHasPermission(scientificInternship))
+			{
+				return Forbid();
+			}
 
-			return View(new ScientificInternshipEditModel(scientificInternship));
+			return View(new ScientificInternshipEditModel(scientificInternship)
+			{
+				Users = _scientificInternshipService.GetUsers(scientificInternship.Id),
+				AllUsers = _userProfileService.GetAll()
+			});
 		}
 
 		// POST: ScientificInternship/Edit/{id}
@@ -85,8 +117,16 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
+			var scientificInternship = _scientificInternshipService.GetById(id);
+			if (!UserHasPermission(scientificInternship))
+			{
+				return Forbid();
+			}
+
 			if (!ModelState.IsValid)
 			{
+				model.Users = _scientificInternshipService.GetUsers(scientificInternship.Id);
+				model.AllUsers = _userProfileService.GetAll();
 				return View(model);
 			}
 
@@ -107,6 +147,11 @@ namespace ScientificReport.Controllers
 			{
 				return NotFound();
 			}
+			
+			if (!UserHasPermission(scientificInternship))
+			{
+				return Forbid();
+			}
 
 			return View(scientificInternship);
 		}
@@ -120,9 +165,92 @@ namespace ScientificReport.Controllers
 			{
 				return NotFound();
 			}
+			
+			if (!UserHasPermission(_scientificInternshipService.GetById(id)))
+			{
+				return Forbid();
+			}
 
 			_scientificInternshipService.DeleteById(id);
 			return RedirectToAction(nameof(Index));
+		}
+		
+		// POST: ScientificInternship/AddUser/{scientificInternshipId}
+		[HttpPost]
+		public IActionResult AddUser(Guid? id, [FromBody] UpdateUserRequest request)
+		{
+			if (id == null)
+			{
+				return NotFound();
+			}
+			
+			var user = _userProfileService.GetById(request.UserId);
+			if (user == null)
+			{
+				return Json(ApiResponse.Fail);
+			}
+			
+			var scientificInternship = _scientificInternshipService.GetById(id.Value);
+			if (scientificInternship == null)
+			{
+				return NotFound();
+			}
+			
+			if (!UserHasPermission(scientificInternship))
+			{
+				return Json(ApiResponse.Fail);
+			}
+			
+			if (!_scientificInternshipService.GetUsers(scientificInternship.Id).Contains(user))
+			{
+				_scientificInternshipService.AddUser(scientificInternship, user);
+			}
+
+			return Json(ApiResponse.Ok);
+		}
+
+		// POST: ScientificInternship/RemoveUser/{scientificInternshipId}
+		[HttpPost]
+		public IActionResult RemoveUser(Guid? id, [FromBody] UpdateUserRequest request)
+		{
+			if (id == null)
+			{
+				return NotFound();
+			}
+			
+			var user = _userProfileService.GetById(request.UserId);
+			if (user == null)
+			{
+				return Json(ApiResponse.Fail);
+			}
+			
+			var publication = _scientificInternshipService.GetById(id.Value);
+			if (publication == null)
+			{
+				return NotFound();
+			}
+			
+			if (!UserHasPermission(publication))
+			{
+				return Json(ApiResponse.Fail);
+			}
+			
+			if (_scientificInternshipService.GetUsers(publication.Id).Contains(user))
+			{
+				_scientificInternshipService.RemoveUser(publication, user);
+			}
+
+			return Json(ApiResponse.Ok);
+		}
+		
+		private bool UserHasPermission(ScientificInternship scientificInternship)
+		{
+			var user = _userProfileService.Get(User);
+			var department = _departmentService.Get(d => d.Staff.Contains(user));
+			return PageHelpers.IsAdmin(User) ||
+				   PageHelpers.IsHeadOfDepartment(User) &&
+				   scientificInternship.UserProfilesScientificInternships.Any(p => department.Staff.Contains(p.UserProfile)) ||
+				   scientificInternship.UserProfilesScientificInternships.Any(p => p.UserProfile.Id == user.Id);
 		}
 	}
 }
